Add context that storage products
