+++ src/Components/ProductsContext/ProductsContext.js
@@ -1,0 +1,3 @@
+import {createContext} from 'react'
+
+export const ProductsContext = createContext(null);
