+++ src/index.js
@@ -2,7 +2,10 @@ import React from 'react'
 import ReactDOM from 'react-dom';
 import MyApp from './MyApp';
 
+
 const divRoot = document.querySelector('#root');
 
-ReactDOM.render( <MyApp />, divRoot);
+
+
+ReactDOM.render(<MyApp />, divRoot);
